+++ lib/core/constants/app_constants.dart
@@ -5,4 +5,5 @@ class AppConstants {
   static const String usersCollection = 'users';
   static const String shiftsCollection = 'shifts';
   static const String tasksCollection = 'tasks';
+  static const String branchesCollection = 'branches';
 }